refactor(exports): reorganize proxy and image module exports
