--- packages/rari/src/proxy/index.ts
@@ -1,16 +1,0 @@
-export { extractParams, matchesPattern, shouldRunProxy } from './matcher'
-export { RariRequest } from './RariRequest'
-export { RariResponse } from './RariResponse'
-export { initializeProxyExecutor } from './runtime-executor'
-export type {
-  CookieOptions,
-  ProxyConfig,
-  ProxyFunction,
-  ProxyMatcher,
-  ProxyModule,
-  ProxyResult,
-  RariFetchEvent,
-  RariURL,
-  RequestCookies,
-  ResponseCookies,
-} from './types'
